refactor(naming): :recycle: StatPage -> HomePage
