+++ app/src/pages/HomePage/contents/AverageCommentLength.tsx
