+++ app/src/pages/HomePage/contents/AverageFeedbackLength.tsx
