+++ app/src/pages/HomePage/contents/CoalitionScoreDynamic.tsx
