+++ app/src/pages/HomePage/contents/CoalitionScoreSum.tsx
