+++ app/src/pages/HomePage/contents/CurrRegisteredCntRank.tsx
