+++ app/src/pages/HomePage/contents/ProjectInfo/ProjectSearchBar.tsx
@@ -6,6 +6,7 @@ import { useDebounce } from '@/utils/useDebounce';
 import { useLazyQuery } from '@apollo/client';
 import styled from '@emotion/styled';
 import { BiSearch } from '@react-icons/all-files/bi/BiSearch';
+import { rgba } from 'emotion-rgba';
 import { useEffect, useRef, useState } from 'react';
 
 const FIND_PROJECT_PREVIEW = gql(/* GraphQL */ `
@@ -25,6 +26,7 @@ type ProjectSearchBarProps = {
 // TODO: Input 로직은 분리할 수 있을 것 같은데...
 export const ProjectSearchBar = ({ onSubmit }: ProjectSearchBarProps) => {
   const [input, setInput] = useState<string>('');
+  const [isFocused, setIsFocused] = useState<boolean>(false);
   const debouncedInput = useDebounce(input, 100);
   const [preview, { loading, error, data }] =
     useLazyQuery(FIND_PROJECT_PREVIEW);
@@ -56,13 +58,15 @@ export const ProjectSearchBar = ({ onSubmit }: ProjectSearchBarProps) => {
 
   return (
     <div css={{ position: 'relative' }}>
-      <ProjectSearchBarLayout>
+      <ProjectSearchBarLayout isFocused={isFocused}>
         <VStack>
           <HStack w="100%" spacing="1rem">
             <Input
               ref={inputRef}
               onChange={(e) => setInput(e.target.value)}
               onKeyDown={handleKeyDown}
+              onFocus={() => setIsFocused(true)}
+              onBlur={() => setIsFocused(false)}
             />
             <Clickable
               onClick={() => handleSubmit(input)}
@@ -92,11 +96,21 @@ export const ProjectSearchBar = ({ onSubmit }: ProjectSearchBarProps) => {
   );
 };
 
-const ProjectSearchBarLayout = styled.div`
-  padding: 1.5rem 2rem 1.5rem 4rem;
-  width: 30rem;
+const ProjectSearchBarLayout = styled.div<{ isFocused: boolean }>`
+  padding: 1rem 2rem 1rem 4rem;
+  width: 24rem;
   border-radius: 3rem;
-  background-color: ${({ theme }) => theme.colors.primary.light};
+  transition: all 0.2s;
+
+  box-shadow: ${({ theme, isFocused }) =>
+    isFocused
+      ? `0 0.4rem 0.4rem ${rgba(theme.colors.mono.black, 0.25)}`
+      : `0 0.4rem 0.4rem ${rgba(theme.colors.mono.black, 0.1)}`};
+
+  &:hover {
+    box-shadow: ${({ theme }) =>
+      `0 0.4rem 0.4rem ${rgba(theme.colors.mono.black, 0.25)}`};
+  }
 `;
 
 const ProjectSearchResult = styled.div`
@@ -108,5 +122,5 @@ const ProjectSearchResult = styled.div`
   padding: 1.5rem 4rem;
   border-radius: 3rem;
   box-shadow: 0 0.4rem 0.4rem rgba(0, 0, 0, 0.25);
-  background-color: ${({ theme }) => theme.colors.primary.light};
+  background-color: ${({ theme }) => theme.colors.mono.white};
 `;
