+++ app/src/pages/HomePage/contents/ProjectInfo/index.ts
