+++ app/src/pages/SkeletonPages/HomePageSkeleton/HomePageSkeleton.tsx
@@ -1,0 +1,6 @@
+import { DashboardSkeleton } from '@/components/templates/DashboardSkeleton';
+import { useHomePageDashboardSkeleton } from './hooks';
+
+export const HomePageSkeleton = () => {
+  return <DashboardSkeleton {...useHomePageDashboardSkeleton()} />;
+};
