+++ app/src/pages/SkeletonPages/HomePageSkeleton/hooks/useHomePageDashboardSkeleton.ts
@@ -1,0 +1,11 @@
+import {
+  homePageDesktopDashboardRows,
+  homePageMobileDashboardRows,
+  homePageTabletDashboardRows,
+} from '@/pages/HomePage/hooks/homePageDashboardRows';
+
+export const useHomePageDashboardSkeleton = () => ({
+  desktopRows: homePageDesktopDashboardRows,
+  tabletRows: homePageTabletDashboardRows,
+  mobileRows: homePageMobileDashboardRows,
+});
