+++ app/src/pages/SkeletonPages/HomePageSkeleton/index.ts
@@ -1,0 +1,1 @@
+export * from './HomePageSkeleton';
