--- app/src/pages/SkeletonPages/StatPageSkeleton/StatPageSkeleton.tsx
@@ -1,6 +1,0 @@
-import { DashboardSkeleton } from '@/components/templates/DashboardSkeleton';
-import { useStatPageDashboardSkeleton } from './hooks';
-
-export const StatPageSkeleton = () => {
-  return <DashboardSkeleton {...useStatPageDashboardSkeleton()} />;
-};
