--- app/src/pages/SkeletonPages/StatPageSkeleton/hooks/index.ts
@@ -1,1 +1,0 @@
-export * from './useStatPageDashboardSkeleton';
