--- app/src/pages/SkeletonPages/StatPageSkeleton/hooks/useStatPageDashboardSkeleton.ts
@@ -1,11 +1,0 @@
-import {
-  statPageDesktopDashboardRows,
-  statPageMobileDashboardRows,
-  statPageTabletDashboardRows,
-} from '@/pages/StatPage/hooks/statPageDashboardRows';
-
-export const useStatPageDashboardSkeleton = () => ({
-  desktopRows: statPageDesktopDashboardRows,
-  tabletRows: statPageTabletDashboardRows,
-  mobileRows: statPageMobileDashboardRows,
-});
