--- app/src/pages/SkeletonPages/StatPageSkeleton/index.ts
@@ -1,1 +1,0 @@
-export * from './StatPageSkeleton';
