--- app/src/pages/StatPage/hooks/index.ts
@@ -1,1 +1,0 @@
-export * from './useStatPageDashboard';
